Fix compiler error

+++ src/TraderXTP/TraderXTP.cpp
@@ -492,7 +492,7 @@ void TraderXTP::OnQueryPosition(XTPQueryStkPositionRsp *position, XTPRI *error_i
 		if (contract)
 		{
 			WTSCommodityInfo* commInfo = contract->getCommInfo();
-			std::string key = fmt::format("{}-{}", code.c_str(), position->position_direction);
+			std::string key = fmt::format("{}-{}", code.c_str(), static_cast<int>(position->position_direction));
 			WTSPositionItem* pos = (WTSPositionItem*)_positions->get(key);
 			if (pos == NULL)
 			{
