+++ src/WtDataStorage/WtRdmDtReader.cpp
@@ -1782,7 +1782,7 @@ WTSKlineSlice* WtRdmDtReader::readKlineSliceByRange(const char* stdCode, WTSKlin
 	WTSCommodityInfo* commInfo = _base_data_mgr->getCommodity(cInfo._exchg, cInfo._product);
 	const char* stdPID = commInfo->getFullPid();
 
-	std::string key = fmt::format("{}#{}", stdCode, period);
+	std::string key = fmt::format("{}#{}", stdCode, static_cast<int>(period));
 	auto it = _bars_cache.find(key);
 	bool bHasHisData = false;
 	if (it == _bars_cache.end())
@@ -2213,7 +2213,7 @@ WTSKlineSlice* WtRdmDtReader::readKlineSliceByCount(const char* stdCode, WTSKlin
 	WTSCommodityInfo* commInfo = _base_data_mgr->getCommodity(cInfo._exchg, cInfo._product);
 	const char* stdPID = commInfo->getFullPid();
 
-	std::string key = fmtutil::format("{}#{}", stdCode, period);
+	std::string key = fmtutil::format("{}#{}", stdCode, static_cast<int>(period));
 	auto it = _bars_cache.find(key);
 	bool bHasHisData = false;
 	if (it == _bars_cache.end())
